tutorial/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
coverage==3.6
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.7
django-heroku==0.3.1
djangorestframework==3.11.0
filelock==3.0.12
gunicorn==20.0.4
httpie==2.2.0
idna==2.10
pipenv==2020.6.2
psycopg2==2.8.5
Pygments==2.6.1
pytz==2020.1
requests==2.24.0
selenium==3.141.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
virtualenv==20.0.25
virtualenv-clone==0.5.4
whitenoise==5.1.0